Break BetaFunctionsProcessor into separate pages
The BetaFunctions pages were fragile because they hard-coded a function
name (as a string) for each possible action, and those strings had to
match identically between display and processor pages.

The best solution seemed to be to break up the BetaFunctionsProcessor
page into separate pages for each function, grouped in a BetaFunctions
namespace. This solution has a bit of extra boilerplate for each new
Page file, but it is by far the most robust. (I had considered making
a BetaFunctions enum class to replace the `func` string, but this was
still too roundabout of a solution.)

Note: This also fixes an error where trying to set political relations
silently did nothing, since we didn't validate that they `func` string
actually matched anything, and the string was hard-coded differently
between the display and processor pages (`Race_relations` and
`Race_Relations` respectively). This error was introduced in e280e9ad9.
