+++ src/pages/Player/BetaFunctions/AddMoneyProcessor.php
@@ -1,0 +1,13 @@
+<?php declare(strict_types=1);
+
+namespace Smr\Pages\Player\BetaFunctions;
+
+use AbstractSmrPlayer;
+
+class AddMoneyProcessor extends BetaFunctionsPageProcessor {
+
+	public function buildBetaFunctionsProcessor(AbstractSmrPlayer $player): void {
+		$player->setCredits(50000000);
+	}
+
+}
