+++ src/pages/Player/BetaFunctions/RemoveWeaponsProcessor.php
@@ -1,0 +1,14 @@
+<?php declare(strict_types=1);
+
+namespace Smr\Pages\Player\BetaFunctions;
+
+use AbstractSmrPlayer;
+
+class RemoveWeaponsProcessor extends BetaFunctionsPageProcessor {
+
+	public function buildBetaFunctionsProcessor(AbstractSmrPlayer $player): void {
+		$ship = $player->getShip();
+		$ship->removeAllWeapons();
+	}
+
+}
