+++ src/pages/Player/BetaFunctions/RepairShipProcessor.php
@@ -1,0 +1,14 @@
+<?php declare(strict_types=1);
+
+namespace Smr\Pages\Player\BetaFunctions;
+
+use AbstractSmrPlayer;
+
+class RepairShipProcessor extends BetaFunctionsPageProcessor {
+
+	public function buildBetaFunctionsProcessor(AbstractSmrPlayer $player): void {
+		$ship = $player->getShip();
+		$ship->setHardwareToMax();
+	}
+
+}
