+++ src/pages/Player/BetaFunctions/SetPoliticalRelationsProcessor.php
@@ -1,0 +1,22 @@
+<?php declare(strict_types=1);
+
+namespace Smr\Pages\Player\BetaFunctions;
+
+use AbstractSmrPlayer;
+use Smr\Database;
+use Smr\Request;
+
+class SetPoliticalRelationsProcessor extends BetaFunctionsPageProcessor {
+
+	public function buildBetaFunctionsProcessor(AbstractSmrPlayer $player): void {
+		$amount = Request::getInt('amount');
+		$race = Request::getInt('race');
+		if ($player->getRaceID() == $race) {
+			create_error('You cannot change race relations with your own race.');
+		}
+		$db = Database::getInstance();
+		$db->write('UPDATE race_has_relation SET relation = ' . $db->escapeNumber($amount) . ' WHERE race_id_1 = ' . $db->escapeNumber($player->getRaceID()) . ' AND race_id_2 = ' . $db->escapeNumber($race) . ' AND game_id = ' . $db->escapeNumber($player->getGameID()));
+		$db->write('UPDATE race_has_relation SET relation = ' . $db->escapeNumber($amount) . ' WHERE race_id_1 = ' . $db->escapeNumber($race) . ' AND race_id_2 = ' . $db->escapeNumber($player->getRaceID()) . ' AND game_id = ' . $db->escapeNumber($player->getGameID()));
+	}
+
+}
