+++ src/pages/Player/BetaFunctions/SetRaceProcessor.php
@@ -1,0 +1,15 @@
+<?php declare(strict_types=1);
+
+namespace Smr\Pages\Player\BetaFunctions;
+
+use AbstractSmrPlayer;
+use Smr\Request;
+
+class SetRaceProcessor extends BetaFunctionsPageProcessor {
+
+	public function buildBetaFunctionsProcessor(AbstractSmrPlayer $player): void {
+		$race = Request::getInt('race');
+		$player->setRaceID($race);
+	}
+
+}
