+++ src/pages/Player/BetaFunctions/SetTurnsProcessor.php
@@ -1,0 +1,14 @@
+<?php declare(strict_types=1);
+
+namespace Smr\Pages\Player\BetaFunctions;
+
+use AbstractSmrPlayer;
+use Smr\Request;
+
+class SetTurnsProcessor extends BetaFunctionsPageProcessor {
+
+	public function buildBetaFunctionsProcessor(AbstractSmrPlayer $player): void {
+		$player->setTurns(Request::getInt('turns'));
+	}
+
+}
